+++ tests/de/aidger/model/budgets/BudgetCheckerTest.java
@@ -89,7 +89,7 @@ public class BudgetCheckerTest {
         contract.save();
 
         course = new Course();
-        course.setId(1);
+        course.setId((long) 1);
         course.setAdvisor("Tester");
         course.setDescription("Description");
         course.setFinancialCategoryId(fc.getId());
@@ -105,7 +105,7 @@ public class BudgetCheckerTest {
         course.save();
 
         employment = new Employment();
-        employment.setId(1);
+        employment.setId((long) 1);
         employment.setAssistantId(assistant.getId());
         employment.setContractId(contract.getId());
         employment.setCourseId(course.getId());
