+++ tests/de/aidger/model/budgets/BudgetCreatorTest.java
@@ -92,7 +92,7 @@ public class BudgetCreatorTest {
         contract.save();
 
         course = new Course();
-        course.setId(1);
+        course.setId((long) 1);
         course.setAdvisor("Tester");
         course.setDescription("Description");
         course.setFinancialCategoryId(fc.getId());
@@ -108,7 +108,7 @@ public class BudgetCreatorTest {
         course.save();
 
         employment = new Employment();
-        employment.setId(1);
+        employment.setId((long) 1);
         employment.setAssistantId(assistant.getId());
         employment.setContractId(contract.getId());
         employment.setCourseId(course.getId());
